refactored validator and removed unnecessary code
