+++ Makefile
@@ -8,7 +8,11 @@ LDFLAGS := -ldl -lglfw -pthread -lm -fsanitize=address -flto -framework Cocoa -f
 HEADERS := -I ./include -I ${BASELIB}/include -I $(LIBMLX)/include
 
 SRC_DEMO := map_mocks.c
-SRC_PARSER := parser/validator.c parser/map_init.c parser/player.c
+SRC_PARSER := parser/validator/validator.c \
+	parser/validator/boundary.c \
+	parser/validator/utils.c \
+	parser/map_init.c parser/player.c
+
 SRC_ENGINE := engine/caster/cast.c
 SRC_MAIN := main.c init.c
 SRCS := $(SRC_MAIN) $(SRC_PARSER) $(SRC_ENGINE) $(SRC_DEMO)
