+++ src/parser/validator/utils.c
@@ -1,0 +1,15 @@
+#include "cube.h"
+
+int	largest_of_three(int a, int b, int c)
+{
+	if (a > b)
+	{
+		if (a > c)
+			return (a);
+		return (c);
+	}
+	if (b > c)
+		return (b);
+	return (c);
+}
+
